64bit ids (WIP)

+++ src/redisearch.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-typedef uint32_t t_docId;
-typedef uint32_t t_offset;
+typedef uint64_t t_docId;
+typedef uint64_t t_offset;
 // used to represent the id of a single field.
 // to produce a field mask we calculate 2^fieldId
 typedef uint16_t t_fieldId;
